requirements.txt
# Веб-фреймворк и сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Валидация и утилиты
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0
python-multipart==0.0.6